Adds concrete service methods for addEvent & addEvents
This removes the need for a user to pass their event data nested inside
an array with a ‘data’ key.

Renaming `data` to `request_body` in the service description. In
hindsight, `data` was too common of a property name and makes it very
hard to support BC issues with this change.