Fix the client when used with guzzle 7.9
Guzzle 7.9 introduces stricter checks on the protocol version. This
protocol version is configured through the `version` config option.
As the whole KeenIOClient config was passed to the instantiated Guzzle
Client, this broke things because the version of the keen service was
passed to Guzzle as the version config as well, while totally unrelated.
The updated logic removes all config options corresponding to the
KeenIOClient from the options passed to the Guzzle client.